TANGO: Fix some compiler warnings
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/general/qmesydaq/+/23392
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>

+++ interfaces/tango/detectorchannel/DetectorChannel.cpp
@@ -708,7 +708,7 @@ Tango::DevBoolean DetectorChannel::update_properties(const Tango::DevVarStringAr
 
 std::string DetectorChannel::incNumber(const std::string &val)
 {
-	int nwidth(5);
+	size_t nwidth(5);
 	std::string tmpString = val;
 	std::string baseName = basename(const_cast<char *>(tmpString.c_str()));
 	std::string::size_type pos = baseName.rfind(".");
